refactor: move CheckRadioProps.ts
